Reworked PVWatts to use pvwatts2014 inputs and simulation module added documentation in sam-arch-3-how-to about table:field variables fixed crash when copying tables: VarTable needed explicit copy constructor to do a deep copy of fields

+++ src/defaultsmanager.cpp
@@ -190,7 +190,7 @@ DefaultsManager::DefaultsManager(wxWindow *parent, int id)
 		}
 	}
 */
-	lblAllConfigurations->SetLabel( wxString::Format("%d total cfgs, %d def. files missing", configs.Count(), configs.Count()-num_defaults) ); 
+	lblAllConfigurations->SetLabel( wxString::Format("%d total cfgs, %d def. files missing", (int)configs.Count(), (int)(configs.Count()-num_defaults)) ); 
 
 	cklConfigurations->Append( configs );
